backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
requests==2.28.1
python-dotenv==0.21.0
pyRFC3339==1.1
google-cloud-bigquery==3.3.3
celery==5.2.7
django-allauth==0.51.0